+++ CHANGES.rst
@@ -39,6 +39,7 @@ Specviz
 Specviz2d
 ^^^^^^^^^
 
+
 API Changes
 -----------
 
@@ -132,6 +133,7 @@ Specviz
 
 Specviz2d
 ^^^^^^^^^
+- Add option to use self-derived spatial profile for Horne extract in spectral extraction plugin. [#2845]
 
 API Changes
 -----------